fix lỗi không chay đc phân ethernet và lỗi khi hiện những chũ dầu
stm32f4_discovery_tsc2046.c
main.c

+++ STM32F4xx_LCD/Project/FreeRTOS/http_client_server_socket/src/stm32f4xx_uartstdio.c
@@ -635,7 +635,7 @@ again:
                 }
 								case 'f':
 								{
-									fValue = va_arg(vaArgP, float);
+									fValue = va_arg(vaArgP, double);
 									ulValue = (long) fValue;
 									ucStart =1;
 									goto convertfloat;
